+++ src/main/java/com/reremouse/lab/math/data/ColumnType.java
@@ -5,5 +5,5 @@ package com.reremouse.lab.math.data;
  * @author lteb2
  */
 public enum ColumnType {
-    String,Float
+    String,Numeric
 }
