+++ src/main/java/com/reremouse/lab/math/ml/lr/IRegression.java
@@ -22,6 +22,6 @@ public interface IRegression {
      * @param x
      * @return 
      */
-    public float predict(IVector x);
+    public double predict(IVector x);
     
 }
