+++ src/main/java/com/reremouse/lab/math/stats/anova/ANOVAResult.java
@@ -7,11 +7,11 @@ package com.reremouse.lab.math.stats.anova;
  */
 public class ANOVAResult {
 
-    float ssBetween;    // 组间平方和（Sum of Squares Between）
-    float ssWithin;     // 组内平方和（Sum of Squares Within）
-    float ssTotal;      // 总平方和（Sum of Squares Total）
-    float fStatistic;   // F统计量
-    float pValue;       // p值
+    double ssBetween;    // 组间平方和（Sum of Squares Between）
+    double ssWithin;     // 组内平方和（Sum of Squares Within）
+    double ssTotal;      // 总平方和（Sum of Squares Total）
+    double fStatistic;   // F统计量
+    double pValue;       // p值
 
     // 构造函数
     // 参数说明：
@@ -20,7 +20,7 @@ public class ANOVAResult {
     //   - ssTotal: 总平方和，等于ssBetween + ssWithin
     //   - fStatistic: F统计量，用于假设检验
     //   - pValue: p值，表示在原假设为真时观察到当前F值或更极端值的概率
-    ANOVAResult(float ssBetween, float ssWithin, float ssTotal, float fStatistic, float pValue) {
+    ANOVAResult(double ssBetween, double ssWithin, double ssTotal, double fStatistic, double pValue) {
         this.ssBetween = ssBetween;
         this.ssWithin = ssWithin;
         this.ssTotal = ssTotal;
