+++ src/main/java/com/reremouse/lab/math/stats/anova/RepeatedMeasuresANOVAResult.java
@@ -7,10 +7,10 @@ package com.reremouse.lab.math.stats.anova;
  */
 public class RepeatedMeasuresANOVAResult {
 
-    float timeF;        // 时间效应的F统计量
-    float timeP;        // 时间效应的p值
-    float subjectF;     // 被试效应的F统计量
-    float subjectP;     // 被试效应的p值
+    double timeF;        // 时间效应的F统计量
+    double timeP;        // 时间效应的p值
+    double subjectF;     // 被试效应的F统计量
+    double subjectP;     // 被试效应的p值
 
     // 构造函数
     // 参数说明：
@@ -18,7 +18,7 @@ public class RepeatedMeasuresANOVAResult {
     //   - timeP: 时间效应的p值，表示时间效应的显著性
     //   - subjectF: 被试效应的F统计量，用于检验被试间差异
     //   - subjectP: 被试效应的p值，表示被试间差异的显著性
-    RepeatedMeasuresANOVAResult(float timeF, float timeP, float subjectF, float subjectP) {
+    RepeatedMeasuresANOVAResult(double timeF, double timeP, double subjectF, double subjectP) {
         this.timeF = timeF;
         this.timeP = timeP;
         this.subjectF = subjectF;
