+++ src/main/java/com/reremouse/lab/math/stats/anova/TwoWayANOVAResult.java
@@ -7,12 +7,12 @@ package com.reremouse.lab.math.stats.anova;
  */
 public class TwoWayANOVAResult {
 
-    float factorAF;        // 因素A的F统计量
-    float factorAP;        // 因素A的p值
-    float factorBF;        // 因素B的F统计量
-    float factorBP;        // 因素B的p值
-    float interactionF;    // 交互效应的F统计量
-    float interactionP;    // 交互效应的p值
+    double factorAF;        // 因素A的F统计量
+    double factorAP;        // 因素A的p值
+    double factorBF;        // 因素B的F统计量
+    double factorBP;        // 因素B的p值
+    double interactionF;    // 交互效应的F统计量
+    double interactionP;    // 交互效应的p值
 
     // 构造函数
     // 参数说明：
@@ -22,8 +22,8 @@ public class TwoWayANOVAResult {
     //   - factorBP: 因素B的p值，表示因素B主效应的显著性
     //   - interactionF: 交互效应的F统计量，用于检验A×B交互效应
     //   - interactionP: 交互效应的p值，表示交互效应的显著性
-    TwoWayANOVAResult(float factorAF, float factorAP, float factorBF, float factorBP,
-            float interactionF, float interactionP) {
+    TwoWayANOVAResult(double factorAF, double factorAP, double factorBF, double factorBP,
+            double interactionF, double interactionP) {
         this.factorAF = factorAF;
         this.factorAP = factorAP;
         this.factorBF = factorBF;
