+++ src/main/java/com/reremouse/lab/math/stats/testing/TestingResult.java
@@ -9,10 +9,10 @@ import com.reremouse.lab.util.Tuple2;
 public class TestingResult {
 
     public Boolean pass = false;
-    public float p = 0.0f;
-    public Tuple2<Float, Float> criticalInteval;
+    public double p = 0.0f;
+    public Tuple2<Double, Double> criticalInteval;
 
-    public TestingResult(Boolean pass, float p, Tuple2 criticalInteval) {
+    public TestingResult(Boolean pass, double p, Tuple2 criticalInteval) {
         this.pass = pass;
         this.p = p;
         this.criticalInteval = criticalInteval;
