+++ src/test/java/com/reremouse/lab/data/DebugDataFrameTest.java
@@ -21,7 +21,7 @@ public class DebugDataFrameTest {
             // 添加Float类型列 / Add Float type column
             Column floatCol1 = new Column();
             floatCol1.setName("feature1");
-            floatCol1.setColumnType(ColumnType.Float);
+            floatCol1.setColumnType(ColumnType.Numeric);
             List<Object> floatData1 = new ArrayList<>();
             floatData1.add(1.0f);
             floatData1.add(2.0f);
@@ -37,7 +37,7 @@ public class DebugDataFrameTest {
             // 添加另一个Float类型列 / Add another Float type column
             Column floatCol2 = new Column();
             floatCol2.setName("feature2");
-            floatCol2.setColumnType(ColumnType.Float);
+            floatCol2.setColumnType(ColumnType.Numeric);
             List<Object> floatData2 = new ArrayList<>();
             floatData2.add(5.0f);
             floatData2.add(6.0f);
